feat(directory in cli, general params)
The cli has been updated so that the argument image_path
can now point either to a single .tif file or a directory
containing several .tif files. If the path to a directory
is supplied, all .tif images will be processed.

Added argument -d to specify that the output should be dumped
to json in the same directory as the image and with an identifiable
name. A json file is stored for each image.

Added argument -s to suppress output to the terminal. This is
useful if you want to dump the metadata from a directory of many
images to json and don't want all the output in the terminal.

Changed the group_parameters_XXX() functions so that the
manufacturer and current filename are always included in the
"General" group. The software version and final lesn type
are now included as standard in the definition of the Class.